requirements.txt
click
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
Flask-socketio
flask-cors
greenlet
gunicorn
importlib-metadata
itsdangerous
Jinja2
Mako
MarkupSafe
packaging
psycopg2
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug
WTForms
zipp

